third_party/WebKit/LayoutTests/inspector/sources/debugger/scripts-panel-expected.txt

Tests that scripts panel UI elements work as intended.


Running: testInitialLoad
  Dumping ScriptsNavigator sources tab:
    (no domain)
      bar.js
      baz.js
      foo.js
      foobar.js

Running: testReset
  Dumping ScriptsNavigator sources tab:
    (no domain)
Revealing in navigator.
  Dumping ScriptsNavigator sources tab:
    (no domain)
      bar.js
      baz.js
      foo.js
  Dumping ScriptsNavigator sources tab:
  Dumping ScriptsNavigator sources tab:
    (no domain)
      bar.js

Running: testDebuggerUISourceCodeAddedAndRemoved
  Dumping ScriptsNavigator sources tab:
    (no domain)
      foo.js
  Dumping ScriptsNavigator sources tab:
    (no domain)
      foo.js
      source.js
